Added the possibility to use all influxdb columns
With this extension all influxdb columns (beside of
time and sequence_number) will be available in Graphite.
This allows to use timeseries automaticaly calculated by
continuous queries, where the column name is named after
the group by function, e.g. sum, mean.

The column-name is separated by the constant string which is
configurable in the config-file. This string enables the possibitity
to identify leafs and therefore to shortcut in case of rendering a
graph. This may be replaced by caching in the finder.